src/main/java/com/example/seniortask3/service/MedicationService.java
package com.example.seniortask3.service;

import com.example.seniortask3.model.Medication;
import com.example.seniortask3.repository.MedicationRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import java.util.List;

@Service
public class MedicationService {

    private final MedicationRepository medicationRepository;

    @Autowired
    public MedicationService(MedicationRepository medicationRepository) {
        this.medicationRepository = medicationRepository;
    }

    // Service methods for medication-related business logic

    public List<Medication> getAllMedications() {
        return medicationRepository.findAll();
    }

    public Medication getMedicationById(Long id) {
        return medicationRepository.findById(id).orElse(null);
    }

    public Medication saveMedication(Medication medication) {
        return medicationRepository.save(medication);
    }

    public void deleteMedicationById(Long id) {
        medicationRepository.deleteById(id);
    }

    // Add other medication-related service methods as needed
}

